+++ examples/fixedExample.js
@@ -1,9 +1,11 @@
 const { FixedThreadPool } = require('poolifier')
 let resolved = 0
+let poolBusy = 0
 const pool = new FixedThreadPool(15, './yourWorker.js', {
   errorHandler: e => console.error(e),
   onlineHandler: () => console.log('worker is online')
 })
+pool.emitter.on('busy', () => poolBusy++)
 
 const start = Date.now()
 const iterations = 1000
@@ -13,7 +15,8 @@ for (let i = 1; i <= iterations; i++) {
     .then(() => {
       resolved++
       if (resolved === iterations) {
-        return console.log('Time take is ' + (Date.now() - start))
+        console.log('Time taken is ' + (Date.now() - start))
+        return console.log('The pool was busy for ' + poolBusy + ' times')
       }
       return null
     })