+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `full` event to dynamic pool.
+
 ## [2.4.1] - 2023-04-05
 
 ### Changed